Add files via upload
This program is designed to demonstrate swapping the values of two variables, `a` and `b`, provided by the user. Initially, the user is prompted to input the values of `a` and `b`. Then, using a temporary variable `temp`, the values of `a` and `b` are swapped. Finally, the program prints the swapped values of `a` and `b` to show the result of the swap. This allows users to see how variables can be exchanged using a temporary variable in Python.